ai_service/model/tiny/requirements.txt
fastapi
llama-cpp-python
uvicorn[standard]